+++ Source/Alcinoe.FMX.Objects.pas
@@ -2477,6 +2477,7 @@ begin
                         AResourceStream, // const AResourceStream: TStream;
                         AMaskResourceName, // const AMaskResourceName: String;
                         AScale, // const AScale: Single;
+                        TRectF.Empty, // const ASrcRect: TRectF;
                         ABufDrawableRect.Width, ABufDrawableRect.Height, // const W, H: single;
                         AApplyMetadataOrientation, // const AApplyMetadataOrientation: Boolean;
                         AWrapMode, // const AWrapMode: TALImageWrapMode;